requirements.txt
pillow~=12.1.0
vllm~=0.13.0
requests~=2.32.5
transformers~=4.57.3
tqdm~=4.67.1
datasets~=4.4.2
fastapi~=0.128.0
uvicorn~=0.40.0
nltk~=3.9.1
scikit-learn~=1.7.1
torch~=2.9.1
setuptools~=78.1.1
python-multipart~=0.0.20
torchvision~=0.24.0
flowdb~=0.1.3
python-dotenv~=1.2.1